Move some data structures/functions into a separate file
